refactor: step を検知する部分の共通化
